fix: duplicate pure workflow xpert
