+++ packages/contracts/src/ai/xpert.model.ts
@@ -752,32 +752,3 @@ export function createAgentConnections(agent: IXpertAgent, collaborators: IXpert
 
   return connections
 }
-
-export function replaceAgentInDraft(draft: TXpertTeamDraft, key: string, agent: IXpertAgent) {
-  const index = draft.nodes.findIndex((_) => _.type === 'agent' && _.key === key)
-  if (index > -1) {
-    draft.nodes[index] = {
-      ...draft.nodes[index],
-      type: 'agent',
-      key: agent.key,
-      entity: agent
-    }
-  } else {
-    throw new Error(`Can't found agent for key: ${key}`)
-  }
-
-  // Replace agent in connections
-  if (key !== agent.key) {
-    draft.connections.forEach((conn) => {
-      if (conn.from === key) {
-        conn.from = agent.key
-        conn.key = `${conn.from}/${conn.to}`
-      } else if (conn.to === key) {
-        conn.to = agent.key
-        conn.key = `${conn.from}/${conn.to}`
-      }
-    })
-  }
-
-  return draft
-}
